chapter_003/src/test/java/ru/job4j/units/UnitsSortTest.java
package ru.job4j.units;

import org.junit.Test;

import static org.hamcrest.core.Is.is;
import static org.junit.Assert.assertThat;

public class UnitsSortTest {
    @Test
    public void whenAscendingOrder() {
        UnitsSort sorter = new UnitsSort();
        String[] in = new String[] {
                "K1\\SK1",
                "K1\\SK2",
                "K1\\SK1\\SSK1",
                "K1\\SK1\\SSK2",
                "K2",
                "K2\\SK1\\SSK1",
                "K2\\SK1\\SSK2"
        };
        String[] expected = new String[] {
                "K1",
                "K1\\SK1",
                "K1\\SK1\\SSK1",
                "K1\\SK1\\SSK2",
                "K1\\SK2",
                "K2",
                "K2\\SK1",
                "K2\\SK1\\SSK1",
                "K2\\SK1\\SSK2"
        };
        String[] out = sorter.sort(in);
        assertThat(out, is(expected));
    }
}
